+++ changelog.md
@@ -4,6 +4,7 @@
 
 ### Fixes
 
+- Change vuex generation to use a default TS client path.
 - Fix cli action org in templates.
 - Seal the capability keeper in the `app.go` template